+ Filling values in navigation drawer; + NavigationDrawerBuilder as singleton + Getting some values from db

+++ app/src/main/java/pl/piotrstaniow/organizeme/NavigationDrawer/NavigationDrawerItem.java
@@ -8,7 +8,7 @@ import android.graphics.drawable.Drawable;
 public class NavigationDrawerItem {
     public Integer icon;
     public CharSequence text;
-    public Integer value;
+    public long value;
     private boolean hasValueState;
 
     public NavigationDrawerItem(Integer icon, CharSequence text) {
